# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BaseRouteResponse": {
        "additionalProperties": false,
        "properties": {
          "distance_meters": {
            "minimum": 0.0,
            "title": "Distance Meters",
            "type": "number"
          },
          "duration_seconds": {
            "minimum": 0.0,
            "title": "Duration Seconds",
            "type": "number"
          },
          "geometry": {
            "$ref": "#/components/schemas/RouteGeometry"
          },
          "maneuvers": {
            "items": {
              "$ref": "#/components/schemas/ManeuverResponse"
            },
            "title": "Maneuvers",
            "type": "array"
          },
          "navigation": {
            "$ref": "#/components/schemas/NavigationTimingResponse"
          },
          "provider": {
            "const": "valhalla",
            "title": "Provider",
            "type": "string"
          },
          "speed_limit_source": {
            "anyOf": [
              {
                "const": "valhalla_graph",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Limit Source"
          },
          "speed_limits": {
            "items": {
              "$ref": "#/components/schemas/RouteSpeedLimitResponse"
            },
            "title": "Speed Limits",
            "type": "array"
          }
        },
        "required": [
          "distance_meters",
          "duration_seconds",
          "geometry",
          "maneuvers",
          "speed_limits",
          "speed_limit_source",
          "provider",
          "navigation"
        ],
        "title": "BaseRouteResponse",
        "type": "object"
      },
...
      "CoordinateRequest": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CoordinateRequest",
        "type": "object"
      },
...
      "CorridorPolicyResponse": {
        "additionalProperties": false,
        "properties": {
          "cap_applied": {
            "enum": [
              "minimum",
              "maximum",
              "none"
            ],
            "title": "Cap Applied",
            "type": "string"
          },
          "effective_cng_range_km": {
            "exclusiveMinimum": 0.0,
            "title": "Effective Cng Range Km",
            "type": "number"
          },
          "radius_km": {
            "exclusiveMinimum": 0.0,
            "title": "Radius Km",
            "type": "number"
          },
          "range_fraction": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Range Fraction",
            "type": "number"
          },
          "uncapped_radius_km": {
            "exclusiveMinimum": 0.0,
            "title": "Uncapped Radius Km",
            "type": "number"
          }
        },
        "required": [
          "effective_cng_range_km",
          "range_fraction",
          "uncapped_radius_km",
          "radius_km",
          "cap_applied"
        ],
        "title": "CorridorPolicyResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NetworkCostBasisResponse": {
        "additionalProperties": false,
        "properties": {
          "distance_model": {
            "const": "road_network",
            "title": "Distance Model",
            "type": "string"
          },
          "duration_model": {
            "enum": [
              "valhalla_graph_speeds",
              "valhalla_time_dependent_traffic"
            ],
            "title": "Duration Model",
            "type": "string"
          },
          "provider": {
            "const": "valhalla",
            "title": "Provider",
            "type": "string"
          },
          "traffic_aware": {
            "title": "Traffic Aware",
            "type": "boolean"
          },
          "traffic_state": {
            "enum": [
              "not_configured",
              "configured",
              "mock",
              "fresh",
              "stale",
              "unavailable"
            ],
            "title": "Traffic State",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "traffic_state",
          "traffic_aware",
          "duration_model",
          "distance_model"
        ],
        "title": "NetworkCostBasisResponse",
        "type": "object"
      },
...
      "PredictiveCandidatesApiRequest": {
        "additionalProperties": false,
        "properties": {
          "costing": {
            "const": "auto",
            "default": "auto",
            "title": "Costing",
            "type": "string"
          },
          "departure_at": {
            "description": "Timezone-aware departure instant used to calculate station and destination ETAs.",
            "format": "date-time",
            "title": "Departure At",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "effective_cng_range_km": {
            "description": "Usable CNG range after the caller's reserve/safety allowance. It controls spatial corridor width only.",
            "exclusiveMinimum": 0.0,
            "maximum": 2000.0,
            "title": "Effective Cng Range Km",
            "type": "number"
          },
          "estimated_remaining_cng_range_km": {
            "description": "Caller-estimated CNG range remaining at the request origin.",
            "exclusiveMinimum": 0.0,
            "maximum": 2000.0,
            "title": "Estimated Remaining Cng Range Km",
            "type": "number"
          },
          "estimated_remaining_gasoline_range_km": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 2000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Caller-estimated gasoline range remaining at the request origin. Supplying it enables an explicit direct-route fallback only when no complete CNG itinerary exists.",
            "title": "Estimated Remaining Gasoline Range Km"
          },
          "excluded_mimit_station_ids": {
            "description": "Official MIMIT station IDs that must not be considered for this replacement plan.",
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Excluded Mimit Station Ids",
            "type": "array"
          },
          "include_closed": {
            "default": false,
            "description": "Include stations evaluated as closed at ETA with a zero opening score. Unknown opening state remains included regardless of this option.",
            "title": "Include Closed",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z]{2}(-[A-Za-z]{2})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "BCP 47 language tag used for maneuver instructions.",
            "title": "Language"
          },
          "maximum_detour_minutes": {
            "description": "Inclusive maximum added road-network travel time.",
            "maximum": 240.0,
            "minimum": 0.0,
            "title": "Maximum Detour Minutes",
            "type": "number"
          },
          "origin": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "origin_heading_degrees": {
            "anyOf": [
              {
                "exclusiveMaximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional observed travel heading at the origin. Used for a moving off-route recalculation so the new route departs in the current direction.",
            "title": "Origin Heading Degrees"
          },
          "origin_heading_tolerance_degrees": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum angular difference accepted at the routed origin.",
            "title": "Origin Heading Tolerance Degrees"
          },
          "reserve_cng_range_km": {
            "description": "Safety range that must remain on arrival at a suggested station.",
            "exclusiveMaximum": 2000.0,
            "minimum": 0.0,
            "title": "Reserve Cng Range Km",
            "type": "number"
          },
          "reserve_gasoline_range_km": {
            "anyOf": [
              {
                "exclusiveMaximum": 2000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Gasoline safety range that must remain at the destination.",
            "title": "Reserve Gasoline Range Km"
          }
        },
        "required": [
          "origin",
          "destination",
          "departure_at",
          "effective_cng_range_km",
          "maximum_detour_minutes",
          "estimated_remaining_cng_range_km",
          "reserve_cng_range_km"
        ],
        "title": "PredictiveCandidatesApiRequest",
        "type": "object"
      },
      "PredictiveCandidatesResponse": {
        "additionalProperties": false,
        "properties": {
          "base_route": {
            "$ref": "#/components/schemas/BaseRouteResponse"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/PredictiveRankedCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "corridor": {
            "$ref": "#/components/schemas/CorridorPolicyResponse"
          },
          "cost_basis": {
            "$ref": "#/components/schemas/NetworkCostBasisResponse"
          },
          "departure_at": {
            "format": "date-time",
            "title": "Departure At",
            "type": "string"
          },
          "excluded_mimit_station_ids": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Mimit Station Ids",
            "type": "array"
          },
          "gasoline_fallback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GasolineFallbackResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "itinerary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictiveItineraryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "maximum_detour_minutes": {
            "minimum": 0.0,
            "title": "Maximum Detour Minutes",
            "type": "number"
          },
          "network_evaluation": {
            "$ref": "#/components/schemas/NetworkEvaluationMetricsResponse"
          },
          "range_basis": {
            "$ref": "#/components/schemas/PredictiveRangeBasisResponse"
          },
          "ranking_evaluation": {
            "$ref": "#/components/schemas/RankingMetricsResponse"
          },
          "ranking_policy": {
            "$ref": "#/components/schemas/RankingPolicyResponse"
          },
          "reachability_evaluation": {
            "$ref": "#/components/schemas/PredictiveReachabilityMetricsResponse"
          },
          "spatial_pruning": {
            "$ref": "#/components/schemas/SpatialPruningMetricsResponse"
          },
          "stage": {
            "const": "predictive_ranking",
            "default": "predictive_ranking",
            "title": "Stage",
            "type": "string"
          },
          "suggestion_state": {
            "enum": [
              "not_needed",
              "suggested",
              "gasoline_fallback",
              "no_reachable_station",
              "no_eligible_station",
              "no_complete_itinerary"
            ],
            "title": "Suggestion State",
            "type": "string"
          }
        },
        "required": [
          "suggestion_state",
          "departure_at",
          "maximum_detour_minutes",
          "excluded_mimit_station_ids",
          "base_route",
          "corridor",
          "spatial_pruning",
          "cost_basis",
          "network_evaluation",
          "range_basis",
          "reachability_evaluation",
          "ranking_policy",
          "ranking_evaluation",
          "candidates",
          "itinerary",
          "gasoline_fallback"
        ],
        "title": "PredictiveCandidatesResponse",
        "type": "object"
      },
...
      "PredictiveRankedCandidateResponse": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/RankedCandidateResponse"
          },
          "estimated_remaining_range_at_arrival_km": {
            "minimum": 0.0,
            "title": "Estimated Remaining Range At Arrival Km",
            "type": "number"
          },
          "reserve_margin_at_arrival_km": {
            "minimum": 0.0,
            "title": "Reserve Margin At Arrival Km",
            "type": "number"
          }
        },
        "required": [
          "candidate",
          "estimated_remaining_range_at_arrival_km",
          "reserve_margin_at_arrival_km"
        ],
        "title": "PredictiveRankedCandidateResponse",
        "type": "object"
      },
...
      "RouteGeometry": {
        "additionalProperties": false,
        "properties": {
          "encoded_polyline": {
            "minLength": 1,
            "title": "Encoded Polyline",
            "type": "string"
          },
          "format": {
            "const": "polyline6",
            "default": "polyline6",
            "title": "Format",
            "type": "string"
          }
        },
        "required": [
          "encoded_polyline"
        ],
        "title": "RouteGeometry",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OptionalHTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CNG-aware navigation API foundations for Italy.",
    "title": "Compass CNG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cng/predictive-candidates": {
      "post": {
        "operationId": "predictive_candidates_api_v1_cng_predictive_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictiveCandidatesApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictiveCandidatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid range request or no route."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid routing provider response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or routing unavailable."
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Predictive Candidates",
        "tags": [
          "predictive-refuelling"
        ]
      }
    },
...
  }
}